requirements.txt
Django==4.2
djangorestframework==3.15
djangorestframework>=3.14.0
djangorestframework-simplejwt==5.3
django-cors-headers==4.3
mysqlclient==2.2
python-dotenv==1.0
anthropic==0.25
openai==1.30
PyPDF2==3.0
#Pillow==10.3
Pillow==10.4.0
python-dateutil==2.8